pyproject.toml
[project]
name = "python-template"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
requires-python = ">=3.11"
dependencies = [
    "g4f>=0.5.6.3",
    "openai>=1.93.2",
    "pillow==11.0.0",
    "pytelegrambotapi==4.25.0",
    "pytesseract==0.3.13",
    "requests>=2.32.4",
]
